Refactor `Size` property to `decimal` and cleanup
Changed `Size` property type from `double` to `decimal` across the codebase for improved precision. Updated DTOs, ViewModels, and the `Property` entity. Adjusted database schema to use `decimal(18, 2)` for the `Size` column and added a migration for this change.

Replaced `CountAvailableAsync` with `GetAvailableCountAsync` in services and repositories. Removed unused methods and redundant code. Made `Name` and `Description` required in `PropertyTypeViewModel`. Improved validation and error handling in controllers.

Performed general cleanup to enhance maintainability and consistency.

+++ Source/Core/RealEstateApp.Core.Application/DTOs/Api/Properties/PropertyApiDto.cs
@@ -11,7 +11,7 @@ namespace RealEstateApp.Core.Application.DTOs.Api.Properties
         public List<string> Improvements { get; set; } = new();
 
         public decimal Price { get; set; }
-        public double Size { get; set; }
+        public decimal Size { get; set; }
         public int Bedrooms { get; set; }
         public int Bathrooms { get; set; }
         public required string Description { get; set; }
