+++ Source/Core/RealEstateApp.Core.Application/DTOs/Property/SavePropertyDto.cs
@@ -9,7 +9,7 @@ namespace RealEstateApp.Core.Application.DTOs.Property
         public string? Code { get; set; } 
         public decimal Price { get; set; }
         public required string Description { get; set; }
-        public double Size { get; set; }
+        public decimal Size { get; set; }
         public int Bedrooms { get; set; }
         public int Bathrooms { get; set; }
         public string? AgentId { get; set; }
